+++ src/components/NostoProduct/events.ts
